+++ crates/loonfs-core/src/gc.rs
@@ -483,13 +483,13 @@ mod tests {
     use super::*;
     use crate::checkpoint::record::set_checkpoint_record_state;
     use crate::checkpoint::{advance_retention_floor, create_checkpoint};
+    use crate::commit_engine::DirectObjectStorePublisher;
     use crate::namespace::bootstrap::bootstrap_namespace;
     use crate::namespace::delete::delete_namespace;
     use crate::namespace::fork::fork_namespace;
     use crate::options::DeleteNamespaceOptions;
     use crate::path::read::{load_metadata_view, ReadLoadContext};
     use crate::publish::PathMutationIntent;
-    use crate::publisher::DirectObjectStorePublisher;
     use crate::storage::content::store_bytes_as_content;
     use loonfs_api::{CommitId, PutBehavior};
     use loonfs_objectstore::fs::LocalFsStore;
